project/requirements.txt
alabaster==0.7.12
asgiref==3.4.1
Babel==2.9.1
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.3
cryptography==3.4.7
Django==3.2.6
django-appconf==1.0.4
django-cryptography==1.0
djangorestframework==3.12.4
docutils==0.17.1
idna==3.2
imagesize==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
pipdeptree==2.1.0
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
snowballstemmer==2.1.0
Sphinx==4.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
urllib3==1.26.6
